requirements.txt
# Core dependencies
openai-whisper>=20231117
pydub>=0.25.1
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pyaudio>=0.2.13
pywebview>=4.0

# For advanced pronunciation scoring (phoneme analysis)
pronouncing>=0.2.0  # CMU Pronouncing Dictionary wrapper
phonemizer>=3.2.0  # Text to phoneme conversion

# Optional: For pronunciation assessment (if available)
# azure-cognitiveservices-speech>=1.30.0

# For grammar analysis (optional)
# spacy>=3.6.0
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz
